Adding description to db migration file and moving reminder dict to messaging
