--- src/dispatch/incident/reminders.py
@@ -1,6 +1,0 @@
-"""Dict for reminder strings and values. Note values are in hours"""
-reminder_select_values = {
-    "thirty": {"message": "30 minutes", "value": 0.5},
-    "one_hour": {"message": "1 hour", "value": 1},
-    "two_hours": {"message": "2 hours", "value": 2},
-}
