+++ src/dispatch/report/messaging.py
@@ -20,6 +20,13 @@ from .models import Report
 
 log = logging.getLogger(__name__)
 
+"""Dict for reminder strings and values. Note values are in hours"""
+reminder_select_values = {
+    "thirty": {"message": "30 minutes", "value": 0.5},
+    "one_hour": {"message": "1 hour", "value": 1},
+    "two_hours": {"message": "2 hours", "value": 2},
+}
+
 
 def get_report_reminder_settings(report_type: ReportTypes):
     report_reminder_settings_map = {
